Add testing for sorting methods

+++ app/src/main/java/com/example/bofteam24/Sorting/ClassSizeSort.java
@@ -23,7 +23,7 @@ public class ClassSizeSort extends SortingStrategy {
     }
 
     @Override
-    public int calculateCourseScore(CourseRoom course) {
+    public double calculateCourseScore(CourseRoom course) {
         return 0;
     }
